+++ src/screens/UTXOManagement/UTXOLabeling.tsx
@@ -2,9 +2,8 @@ import React, { useContext, useEffect, useRef, useState } from 'react';
 import { useNavigation, useRoute } from '@react-navigation/native';
 import KeeperHeader from 'src/components/KeeperHeader';
 import ScreenWrapper from 'src/components/ScreenWrapper';
-import { StyleSheet, TouchableOpacity, View, ScrollView, Keyboard, Vibration } from 'react-native';
-import { Box, Input, useColorMode } from 'native-base';
-import Buttons from 'src/components/Buttons';
+import { StyleSheet, TouchableOpacity, View, ScrollView } from 'react-native';
+import { Box, useColorMode } from 'native-base';
 import { hp, wp } from 'src/constants/responsive';
 import { UTXO } from 'src/services/wallets/interfaces';
 import { LabelRefType, NetworkType } from 'src/services/wallets/enums';
@@ -16,192 +15,20 @@ import BtcWhite from 'src/assets/images/btc_white.svg';
 import Text from 'src/components/KeeperText';
 import openLink from 'src/utils/OpenLink';
 import config from 'src/utils/service-utilities/config';
-import ConfirmSquare from 'src/assets/images/confirm-square.svg';
-import ConfirmSquareGreen from 'src/assets/images/confirm-square-green.svg';
 import useCurrencyCode from 'src/store/hooks/state-selectors/useCurrencyCode';
 import { useAppSelector } from 'src/store/hooks';
 import useExchangeRates from 'src/hooks/useExchangeRates';
 import { getAmt, getCurrencyImageByRegion, getUnit } from 'src/constants/Bitcoin';
 import useToastMessage from 'src/hooks/useToastMessage';
-import ToastErrorIcon from 'src/assets/images/toast_error.svg';
 import useLabelsNew from 'src/hooks/useLabelsNew';
-import { resetState, setSyncingUTXOError } from 'src/store/reducers/utxos';
-import LabelItem from './components/LabelItem';
-import { Colors } from 'react-native/Libraries/NewAppScreen';
 import Link from 'src/assets/images/link.svg';
 import LinkWhite from 'src/assets/images/link-white.svg';
 import Edit from 'src/assets/images/edit.svg';
 import EditWhite from 'src/assets/images/edit-white.svg';
-import { LocalizationContext, translations } from 'src/context/Localization/LocContext';
+import { LocalizationContext } from 'src/context/Localization/LocContext';
 import { EditNoteContent } from '../ViewTransactions/TransactionDetails';
 import KeeperModal from 'src/components/KeeperModal';
-
-function getLabelChanges(existingLabels, updatedLabels) {
-  const existingNames = new Set(existingLabels.map((label) => label.name));
-  const updatedNames = new Set(updatedLabels.map((label) => label.name));
-
-  const addedLabels = updatedLabels.filter((label) => !existingNames.has(label.name));
-  const deletedLabels = existingLabels.filter((label) => !updatedNames.has(label.name));
-
-  const labelChanges = {
-    added: addedLabels,
-    deleted: deletedLabels,
-  };
-
-  return labelChanges;
-}
-
-export function LabelsEditor({ utxo = null, address = null, wallet, onLabelsSaved }) {
-  const { labels } = useLabelsNew({ address, utxos: utxo ? [utxo] : [], wallet });
-  const { syncingUTXOs, apiError } = useAppSelector((state) => state.utxos);
-  const { showToast } = useToastMessage();
-  const processDispatched = useRef(false);
-  const dispatch = useDispatch();
-  const labelsKey = address ? address : `${utxo.txId}:${utxo.vout}`;
-
-  const getSortedNames = (labels) =>
-    labels
-      .sort((a, b) =>
-        a.isSystem < b.isSystem ? 1 : a.isSystem > b.isSystem ? -1 : a.name > b.name ? 1 : -1
-      )
-      .reduce((a, c) => {
-        a += c.name;
-        return a;
-      }, '');
-
-  const [existingLabels, setExistingLabels] = useState([]);
-  const [label, setLabel] = useState('');
-  const [editingIndex, setEditingIndex] = useState(-1);
-  const { colorMode } = useColorMode();
-  const lablesUpdated = getSortedNames(labels[labelsKey]) !== getSortedNames(existingLabels);
-
-  useEffect(() => {
-    setExistingLabels(labels ? labels[labelsKey] || [] : []);
-    return () => {
-      dispatch(resetState());
-    };
-  }, []);
-
-  useEffect(() => {
-    if (apiError) {
-      showToast(apiError.toString(), <ToastErrorIcon />);
-      dispatch(setSyncingUTXOError(null));
-      processDispatched.current = false;
-    }
-    if (processDispatched.current && !syncingUTXOs) {
-      processDispatched.current = false;
-
-      onLabelsSaved();
-    }
-  }, [apiError, syncingUTXOs]);
-
-  const onCloseClick = (index) => {
-    existingLabels.splice(index, 1);
-    setExistingLabels([...existingLabels]);
-  };
-
-  const onEditClick = (item, index) => {
-    setLabel(item.name);
-    setEditingIndex(index);
-  };
-
-  const onAdd = () => {
-    if (label) {
-      if (editingIndex !== -1) {
-        existingLabels[editingIndex] = { name: label, isSystem: false };
-      } else {
-        existingLabels.push({ name: label, isSystem: false });
-      }
-      setEditingIndex(-1);
-      setExistingLabels(existingLabels);
-      setLabel('');
-      Vibration.vibrate(50);
-      Keyboard.dismiss();
-    }
-  };
-
-  const onSaveChangeClick = async () => {
-    Keyboard.dismiss();
-    const finalLabels = existingLabels.filter(
-      (label) => !label.isSystem // ignore the system label since they are internal references
-    );
-    const initialLabels = labels[labelsKey].filter((label) => !label.isSystem);
-    const labelChanges = getLabelChanges(initialLabels, finalLabels);
-    processDispatched.current = true;
-    if (address) {
-      dispatch(bulkUpdateLabels({ labelChanges, address, wallet }));
-    } else {
-      dispatch(bulkUpdateLabels({ labelChanges, UTXO: utxo, wallet }));
-    }
-  };
-
-  return (
-    <Box>
-      <Box
-        style={[
-          styles.listContainer,
-          colorMode === 'dark' && { borderWidth: 1, borderColor: 'rgba(31, 31, 31, 0.2)' },
-        ]}
-        backgroundColor={`${colorMode}.boxBackground`}
-      >
-        <Box
-          style={styles.inputLabeWrapper}
-          backgroundColor={`${colorMode}.seashellWhite`}
-          borderColor={`${colorMode}.borderBrown`}
-        >
-          <Box style={styles.inputLabelBox}>
-            <Input
-              testID="input_utxoLabel"
-              onChangeText={(text) => {
-                setLabel(text);
-              }}
-              style={styles.inputLabel}
-              height={hp(38)}
-              borderWidth={0}
-              placeholder={`+ ${translations.wallet.AddLabels}`}
-              value={label}
-              autoCorrect={false}
-              backgroundColor={`${colorMode}.seashellWhite`}
-              _input={
-                colorMode === 'dark' && {
-                  selectionColor: Colors.SecondaryWhite,
-                  cursorColor: Colors.SecondaryWhite,
-                }
-              }
-            />
-          </Box>
-          <TouchableOpacity style={styles.addBtnWrapper} onPress={onAdd} testID="btn_addUtxoLabel">
-            {label && label !== '' ? <ConfirmSquareGreen /> : <ConfirmSquare />}
-          </TouchableOpacity>
-        </Box>
-        {existingLabels && existingLabels.length > 0 && (
-          <View style={styles.listSubContainer}>
-            {existingLabels.map((item, index) => (
-              <LabelItem
-                item={item}
-                index={index}
-                key={`${item.name}:${item.isSystem}`}
-                editingIndex={editingIndex}
-                onCloseClick={onCloseClick}
-                onEditClick={onEditClick}
-              />
-            ))}
-          </View>
-        )}
-      </Box>
-      {lablesUpdated && (
-        <Box style={styles.ctaBtnWrapper}>
-          <Buttons
-            primaryLoading={syncingUTXOs}
-            primaryCallback={onSaveChangeClick}
-            primaryText="Save Labels"
-            fullWidth
-          />
-        </Box>
-      )}
-    </Box>
-  );
-}
+import LabelsEditor, { getLabelChanges } from './components/LabelsEditor';
 
 function UTXOLabeling() {
   const { showToast } = useToastMessage();
